WR490876: Fix issues from security audit.
Fixes: S5, S1, S3, S8.

+++ csp_report.php
@@ -32,7 +32,7 @@ if ($viewblockeddomain) {
 }
 $removedirective = optional_param('removedirective', false, PARAM_TEXT);
 $removedomain = optional_param('removedomain', false, PARAM_TEXT);
-$removerecordwithid = optional_param('removerecordwithid', false, PARAM_TEXT);
+$removerecordwithid = optional_param('removerecordwithid', false, PARAM_INT);
 $download = optional_param('download', '', PARAM_ALPHA);
 
 admin_externalpage_setup('local_csp_report', '', null, '', ['pagelayout' => 'report']);
@@ -67,7 +67,7 @@ $PAGE->set_url('/local/csp/csp_report.php', [
 ]);
 
 $resetallcspstatistics = optional_param('resetallcspstatistics', 0, PARAM_INT);
-if ($resetallcspstatistics == 1 && confirm_sesskey()) {
+if ($resetallcspstatistics == 1 && confirm_sesskey() && data_submitted()) {
     $DB->delete_records('local_csp');
     redirect(new moodle_url('/local/csp/csp_report.php'));
 }
